Extract engine merge operations behind the facade
merge and _merge_with_transitions form a coupled operation family with shared concat, transition, duration, and probe behavior. This moves both into engine_merge.py while re-exporting the public and legacy private names from mcp_video.engine.

Constraint: Existing callers import merge from mcp_video.engine, and internal tests may still reach _merge_with_transitions.

Rejected: Split _merge_with_transitions separately | it is private implementation detail used only by merge and would create an artificial boundary.

Confidence: high

Scope-risk: moderate

Directive: Keep engine.py as the compatibility facade while operation families move into focused modules.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_merge.py

Tested: /opt/homebrew/bin/python3 merge/_merge_with_transitions re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'merge' tests/test_server.py -k 'merge' tests/test_e2e.py -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
